Code compiling using the latest dmd compiler (+dub as build tool).

+++ src/ivan/ifs3d/mousestate.d
@@ -1,7 +1,7 @@
 module ivan.ifs3d.mousestate;
 
 class MouseState {
-	int X = 100, Y = 100;
+	double X = 100, Y = 100;
 	int XDelta = 0, YDelta = 0;
 	bool Left = false, Middle = false, Right = false;
 	bool LeftOld = false, MiddleOld = false, RightOld = false;
